eft_havale_app/requirements.txt
Faker==37.4.2
kafka-python==2.2.15
python-dotenv==1.1.1
tzdata==2025.2
